Reorganize demo prefab packages
Changes
- Split space game object classes into asteroids and ships subpackages
