+++ mayonez-demos/src/main/java/slavsquatsuperstar/demos/spacegame/objects/asteroids/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Destructible asteroids that serve as obstacles.
+ *
+ * @author SlavSquatSuperstar
+ */
+package slavsquatsuperstar.demos.spacegame.objects.asteroids;
