+++ mayonez-demos/src/main/java/slavsquatsuperstar/demos/spacegame/objects/ships/EnemyShip.java
@@ -1,10 +1,11 @@
-package slavsquatsuperstar.demos.spacegame.objects;
+package slavsquatsuperstar.demos.spacegame.objects.ships;
 
 import mayonez.math.*;
 import mayonez.physics.*;
 import slavsquatsuperstar.demos.spacegame.combat.EnemyFireController;
 import slavsquatsuperstar.demos.spacegame.movement.EnemyThrustController;
 import slavsquatsuperstar.demos.spacegame.movement.ThrusterPrefabs;
+import slavsquatsuperstar.demos.spacegame.objects.SpawnManager;
 
 /**
  * An enemy spaceship.
