+++ mayonez-demos/src/main/java/slavsquatsuperstar/demos/spacegame/objects/ships/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Player- or computer-controlled spaceships.
+ *
+ * @author SlavSquatSuperstar
+ */
+package slavsquatsuperstar.demos.spacegame.objects.ships;
