+++ preferences.json
@@ -1,6 +1,6 @@
 {
     "title": "Mayonez Engine",
-    "version": "v0.7.10-snapshot",
+    "version": "v0.7.10-pre2-snapshot",
     "screen_width": 1200,
     "screen_height": 800,
     "fps": 60,
